+++ cola/widgets/commitmsg.py
@@ -194,6 +194,25 @@ class CommitMessageEditor(QtWidgets.QFrame):
         self.fixup_commit_menu = self.actions_menu.addMenu(N_('Fixup Previous Commit'))
         self.fixup_commit_menu.aboutToShow.connect(self.build_fixup_menu)
 
+        # Author and Date display.
+        self._author_image = qtutils.pixmap_label(
+            icons.person(), defs.default_icon, parent=self
+        )
+        self._author_label = qtutils.plain_text_label(parent=self)
+        self._date_image = qtutils.pixmap_label(
+            icons.clock(), defs.default_icon, parent=self
+        )
+        self._date_label = qtutils.plain_text_label(parent=self)
+
+        self.bottomlayout = qtutils.hbox(
+            defs.no_margin,
+            defs.spacing,
+            self._author_image,
+            self._author_label,
+            qtutils.STRETCH,
+            self._date_image,
+            self._date_label,
+        )
         self.toplayout = qtutils.hbox(
             defs.no_margin,
             defs.spacing,
@@ -206,7 +225,9 @@ class CommitMessageEditor(QtWidgets.QFrame):
         self.topwidget = QtWidgets.QWidget()
         self.topwidget.setLayout(self.toplayout)
 
-        self.mainlayout = qtutils.vbox(defs.no_margin, defs.spacing, self.description)
+        self.mainlayout = qtutils.vbox(
+            defs.no_margin, defs.spacing, self.description, self.bottomlayout
+        )
         self.setLayout(self.mainlayout)
 
         qtutils.connect_button(self.commit_button, self.commit)
@@ -449,6 +470,42 @@ class CommitMessageEditor(QtWidgets.QFrame):
                 with qtutils.BlockSignals(self.commit_author_action):
                     self.commit_author_action.setChecked(True)
 
+        self.update_author_and_date()
+
+    def update_author_and_date(self):
+        """Hide or display the author field"""
+        amending = self.model.mode == self.model.mode_amend
+        # Display the author when amending or when overridden.
+        if self._git_commit_author:
+            author = self._git_commit_author
+        elif amending:
+            author = self.model.commit_author
+        else:
+            author = ''
+        author_visible = bool(author)
+        if author:
+            self._author_label.setText(author)
+        self._author_image.setVisible(author_visible)
+        self._author_label.setVisible(author_visible)
+
+        # Display the commit date when amending or when overridden.
+        if self._git_commit_date:
+            date = self._git_commit_date
+        elif amending and self._last_commit_datetime:
+            date = display.git_commit_date(self._last_commit_datetime)
+        else:
+            date = ''
+        date_visible = bool(date)
+        if date:
+            self._date_label.setText(date)
+        self._date_image.setVisible(date_visible)
+        self._date_label.setVisible(date_visible)
+
+        if author_visible or date_visible:
+            self.actions_button.setIcon(icons.gear_solid())
+        else:
+            self.actions_button.setIcon(icons.configure())
+
     def commit(self):
         """Attempt to create a commit from the index and commit message."""
         context = self.context
@@ -669,6 +726,7 @@ class CommitMessageEditor(QtWidgets.QFrame):
         """Choose the date and time that is used when authoring commits"""
         if not enabled:
             self._git_commit_date = None
+            self.update_author_and_date()
             return
         widget = CommitDateDialog(
             self, self.context, commit_datetime=self._last_commit_datetime
@@ -680,6 +738,7 @@ class CommitMessageEditor(QtWidgets.QFrame):
             self._last_commit_datetime = CommitDateDialog.tick_time(widget.datetime())
         else:
             self.commit_date_action.setChecked(False)
+        self.update_author_and_date()
 
     def set_commit_author(self, enabled):
         """Choose a commit author to override the author value when authoring commits"""
@@ -687,6 +746,7 @@ class CommitMessageEditor(QtWidgets.QFrame):
             if self._git_commit_author:
                 self._last_git_commit_author = self._git_commit_author
             self._git_commit_author = None
+            self.update_author_and_date()
             return
         widget = CommitAuthorDialog(
             self,
@@ -708,6 +768,7 @@ class CommitMessageEditor(QtWidgets.QFrame):
         else:
             self._git_commit_author = None
             self.commit_author_action.setChecked(False)
+        self.update_author_and_date()
 
     # Qt overrides
     def showEvent(self, event):